requirements.txt
# Core AI Framework
   crewai>=0.80.0
   crewai-tools>=0.12.0
   
   # LLM Integration
   langchain-anthropic>=0.3.0
   anthropic>=0.39.0
   
   # Stock Data
   yfinance>=0.2.0
   
   # Utilities
   python-dotenv>=1.0.0
   pandas>=2.0.0
   
   # Web API
   flask>=3.0.0
   flask-cors>=4.0.0
   gunicorn>=21.2.0